Changed Sliders back to MySliders and on_release instead of on_touch_up. For some reason if the original (not custom) Sliders are used the GUI becomes incredibly slow!
Also added buttons (without functionality) for cutting the song and adding and deleting syllables.

Got rid of right touch so that dots cannot be made on the screen when the right button is clicked.

Currently hacked to just use defined directory/folder instead of the FileBrowser input so that I can troubleshoot more quickly.