Refacotring analysis.
Before, the analysis class took a directory, processed each song(files in directory), then returned results.

I factored out parts of the analysis that deals with the "Song" and created a class for it. Now a song can be processed without having to go through an entire directory. This means we can spawn off processes per song while people are iterate through processing the songs.

Also extracted out functions to be modular. Separates the functions from the class. I did this mainly for understanding the data flow, as well as simplifying structure of code.