Now adding the function of the top of the range slider for a low pass filter.
Ended up just using the entire value of that slider (which is a list) rather than value1 and value2 separately so that the number of parameters does not increase. This is very similar to how I used the crop range slider. This also made it easier to deal with old gzip files when loading parameters from previously chippered files since they would not have a list and only have a single value for high pass filter (no low pass filter).

Still need to change the names from high pass filter or hpf to something general for both low and high pass filtering (in all three files).

Added TODO note to make sure we go back and maybe clean up or optimize how self variables are being used versus the non self versions. (First time using these so not really sure I did so properly.)

Found a bug in the range slider. Value1 can be > value2 which causes problems for both the filtering and cropping. Need to fix this! Added TODO.