Changed names from high pass filter or hpf to something general for both low and high pass filtering (in all three files). Fixed the parameter name in output file and the check for parameters from previously chippered files to reflex the change.
Easy fix to account for bug in range slider: instead of changing the actual slider, I just now check value1 < value2 and if not, then change the order of these in the list. Did this for both filter_boundary and bout_range (since it was causing odd cases in crop as well). Added a TODO to make more permanent in the range slider script instead, just a reminder, not a must.