Added landing page with logo, general information and useful links. This information/links can be changed or updated but the placeholders and spacing is now setup to make this easier.
Added radio buttons for the user to choose what they will do in Chipper - syllable segmentation, determine note and syllable similarity thresholds, and analyze. The appropriate settings are enabled for the user to adjust before starting (not currently working/bound to the following screens).

Start button. Currently this does the same thing no matter which radio button is chosen. The screens for the thresholding and analysis must be added and corresponding logic will be added after the filechooser screen.