Removed user input for low and high pass filter.
Connected percent signal threshold, min silence duration, and min syllable duration from user input to the sliders when chippering. Made sure the input could only be one decimal place (this is also true for percent signal even with change of slider). Zero is not possible for % signal. If zero is input for the durations, it will be rounded to ms equal to one pixel.