Adding log file for error (exception catching) output.
Re-simplified piece of code at the end. I unraveled it to time it  (it was the most expensive part of code) and it was in the multiple flattens. I tried to flatten it once then slice later, but was never able to come up with a method to do so.